LPS-99919 Move classes to portal-cache-ehcache-impl; portal-cache-ehcache-spi was created because of portal-cache-ehcache-multiple, which was removed in LPS-87729
