+++ modules/apps/portal-cache/portal-cache-ehcache-impl/src/main/java/com/liferay/portal/cache/ehcache/internal/EhcacheUnwrapUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.cache.ehcache.spi;
+package com.liferay.portal.cache.ehcache.internal;
 
 import com.liferay.portal.cache.PortalCacheWrapper;
 import com.liferay.portal.kernel.cache.PortalCache;
