+++ modules/apps/portal-cache/portal-cache-ehcache-impl/src/main/java/com/liferay/portal/cache/ehcache/internal/event/EhcachePortalCacheListenerFactory.java
@@ -18,8 +18,6 @@ import com.liferay.portal.cache.PortalCacheListenerFactory;
 import com.liferay.portal.cache.PortalCacheReplicator;
 import com.liferay.portal.cache.PortalCacheReplicatorFactory;
 import com.liferay.portal.cache.ehcache.internal.EhcacheConstants;
-import com.liferay.portal.cache.ehcache.spi.event.ConfigurableEhcachePortalCacheListener;
-import com.liferay.portal.cache.ehcache.spi.event.EhcachePortalCacheListenerAdapter;
 import com.liferay.portal.kernel.cache.PortalCache;
 import com.liferay.portal.kernel.cache.PortalCacheException;
 import com.liferay.portal.kernel.cache.PortalCacheListener;
