+++ modules/apps/portal-cache/portal-cache-ehcache-impl/src/main/java/com/liferay/portal/cache/ehcache/internal/event/EhcachePortalCacheManagerListenerFactory.java
@@ -17,7 +17,6 @@ package com.liferay.portal.cache.ehcache.internal.event;
 import com.liferay.portal.cache.PortalCacheManagerListenerFactory;
 import com.liferay.portal.cache.ehcache.internal.EhcacheConstants;
 import com.liferay.portal.cache.ehcache.internal.EhcachePortalCacheManager;
-import com.liferay.portal.cache.ehcache.spi.event.EhcachePortalCacheManagerListenerAdapter;
 import com.liferay.portal.kernel.cache.PortalCacheManagerListener;
 import com.liferay.portal.kernel.exception.SystemException;
 import com.liferay.portal.kernel.util.InstanceFactory;
